requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.9.0
plotly>=5.10.0
python-dotenv>=0.19.0
requests>=2.28.0
openpyxl>=3.0.0
# Advanced Analytics Requirements
scikit-learn>=1.3.0
statsmodels>=0.14.0
networkx>=3.0
geopandas>=0.13.0
folium>=0.14.0